Add product adding functionality

+++ database/migrations/2020_08_16_183920_create_products_table.php
@@ -27,7 +27,7 @@ class CreateProductsTable extends Migration
             $table->timestamps();
 
             //Исключительно нужно для интеграции с Poster Pos
-            $table->unsignedInteger('poster_id');
+            $table->unsignedBigInteger('poster_id');
 
         });
     }
